+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master Dev Branch
 
 ### 2026-07-24
+- Better filter handling via session
 - Initial Charts support via Chart.js
 
 ### 2026-07-21